feat: rename `SimpleCacheClient` to `CacheClient`
Update code and documentation to refer to the cache client as
`CacheClient` (`CacheClientAsync`) instead of
`SimpleCacheClient` (`SimpleCacheClientAsync`).
