+++ src/momento/__init__.py
@@ -1,10 +1,10 @@
 import logging
 
 from .auth import CredentialProvider
+from .cache_client import CacheClient
+from .cache_client_async import CacheClientAsync
 from .config import Configurations
-from .simple_cache_client import SimpleCacheClient
-from .simple_cache_client_async import SimpleCacheClientAsync
 
 logging.getLogger("momentosdk").addHandler(logging.NullHandler())
 
-__all__ = ["CredentialProvider", "Configurations", "SimpleCacheClient", "SimpleCacheClientAsync"]
+__all__ = ["CredentialProvider", "Configurations", "CacheClient", "CacheClientAsync"]
