+++ src/momento/cache_client.py
@@ -12,8 +12,8 @@ from momento.requests import CollectionTtl
 
 try:
     from momento.internal._utilities import _validate_request_timeout
-    from momento.internal.aio._scs_control_client import _ScsControlClient
-    from momento.internal.aio._scs_data_client import _ScsDataClient
+    from momento.internal.synchronous._scs_control_client import _ScsControlClient
+    from momento.internal.synchronous._scs_data_client import _ScsDataClient
 except ImportError as e:
     if e.name == "cygrpc":
         import sys
@@ -83,8 +83,8 @@ from momento.responses import (
 from momento.typing import TDictionaryItems
 
 
-class SimpleCacheClientAsync:
-    """Async Simple Cache Client.
+class CacheClient:
+    """Synchronous Cache Client.
 
     Cache and control methods return a response object unique to each request.
     The response object is resolved to a type-safe object of one of several
@@ -93,7 +93,7 @@ class SimpleCacheClientAsync:
     Pattern matching can be used to operate on the appropriate subtype.
     For example, in python 3.10+ if you're deleting a key::
 
-        response = await client.delete(cache_name, key)
+        response = client.delete(cache_name, key)
         match response:
             case CacheDelete.Success():
                 ...they key was deleted or not found...
@@ -102,7 +102,7 @@ class SimpleCacheClientAsync:
 
     or equivalently in earlier versions of python::
 
-        response = await client.delete(cache_name, key)
+        response = client.delete(cache_name, key)
         if isinstance(response, CacheDelete.Success):
             ...
         elif isinstance(response, CacheDelete.Error):
@@ -139,12 +139,12 @@ class SimpleCacheClientAsync:
         Example::
 
             from datetime import timedelta
-            from momento import Configurations, CredentialProvider, SimpleCacheClientAsync
+            from momento import Configurations, CredentialProvider, SimpleCacheClient
 
             configuration = Configurations.Laptop.latest()
             credential_provider = CredentialProvider.from_environment_variable("MOMENTO_AUTH_TOKEN")
             ttl_seconds = timedelta(seconds=60)
-            client = SimpleCacheClientAsync(configuration, credential_provider, ttl_seconds)
+            client = SimpleCacheClient(configuration, credential_provider, ttl_seconds)
         """
         _validate_request_timeout(configuration.get_transport_strategy().get_grpc_configuration().get_deadline())
         self._logger = logs.logger
@@ -152,24 +152,23 @@ class SimpleCacheClientAsync:
         self._control_client = _ScsControlClient(configuration, credential_provider)
         self._cache_endpoint = credential_provider.cache_endpoint
         self._data_clients = [
-            _ScsDataClient(configuration, credential_provider, default_ttl)
-            for _ in range(SimpleCacheClientAsync._NUM_CLIENTS)
+            _ScsDataClient(configuration, credential_provider, default_ttl) for _ in range(CacheClient._NUM_CLIENTS)
         ]
 
-    async def __aenter__(self) -> SimpleCacheClientAsync:
+    def __enter__(self) -> CacheClient:
         return self
 
-    async def __aexit__(
+    def __exit__(
         self,
         exc_type: Optional[Type[BaseException]],
         exc_value: Optional[BaseException],
         traceback: Optional[TracebackType],
     ) -> None:
-        await self._control_client.close()
+        self._control_client.close()
         for data_client in self._data_clients:
-            await data_client.close()
+            data_client.close()
 
-    async def create_cache(self, cache_name: str) -> CreateCacheResponse:
+    def create_cache(self, cache_name: str) -> CreateCacheResponse:
         """Creates a cache if it doesn't exist.
 
         Args:
@@ -178,9 +177,9 @@ class SimpleCacheClientAsync:
         Returns:
             CreateCacheResponse:
         """
-        return await self._control_client.create_cache(cache_name)
+        return self._control_client.create_cache(cache_name)
 
-    async def delete_cache(self, cache_name: str) -> DeleteCacheResponse:
+    def delete_cache(self, cache_name: str) -> DeleteCacheResponse:
         """Deletes a cache and all of the items within it.
 
         Args:
@@ -189,17 +188,17 @@ class SimpleCacheClientAsync:
         Returns:
             DeleteCacheResponse:
         """
-        return await self._control_client.delete_cache(cache_name)
+        return self._control_client.delete_cache(cache_name)
 
-    async def list_caches(self) -> ListCachesResponse:
+    def list_caches(self) -> ListCachesResponse:
         """Lists all caches.
 
         Returns:
             ListCachesResponse:
         """
-        return await self._control_client.list_caches()
+        return self._control_client.list_caches()
 
-    async def create_signing_key(self, ttl: timedelta) -> CreateSigningKeyResponse:
+    def create_signing_key(self, ttl: timedelta) -> CreateSigningKeyResponse:
         """Creates a Momento signing key.
 
         Args:
@@ -208,9 +207,9 @@ class SimpleCacheClientAsync:
         Returns:
             CreateSigningKeyResponse
         """
-        return await self._control_client.create_signing_key(ttl, self._cache_endpoint)
+        return self._control_client.create_signing_key(ttl, self._cache_endpoint)
 
-    async def revoke_signing_key(self, key_id: str) -> RevokeSigningKeyResponse:
+    def revoke_signing_key(self, key_id: str) -> RevokeSigningKeyResponse:
         """Revokes a Momento signing key, all tokens signed by which will be invalid.
 
         Args:
@@ -219,17 +218,17 @@ class SimpleCacheClientAsync:
         Returns:
             RevokeSigningKeyResponse
         """
-        return await self._control_client.revoke_signing_key(key_id)
+        return self._control_client.revoke_signing_key(key_id)
 
-    async def list_signing_keys(self) -> ListSigningKeysResponse:
+    def list_signing_keys(self) -> ListSigningKeysResponse:
         """Lists all Momento signing keys for the provided auth token.
 
         Returns:
             ListSigningKeysResponse
         """
-        return await self._control_client.list_signing_keys(self._cache_endpoint)
+        return self._control_client.list_signing_keys(self._cache_endpoint)
 
-    async def increment(
+    def increment(
         self,
         cache_name: str,
         key: str | bytes,
@@ -249,9 +248,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheIncrementResponse
         """
-        return await self._data_client.increment(cache_name, key, amount, ttl)
+        return self._data_client.increment(cache_name, key, amount, ttl)
 
-    async def set(
+    def set(
         self,
         cache_name: str,
         key: str | bytes,
@@ -271,9 +270,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetResponse:
         """
-        return await self._data_client.set(cache_name, key, value, ttl)
+        return self._data_client.set(cache_name, key, value, ttl)
 
-    async def set_if_not_exists(
+    def set_if_not_exists(
         self,
         cache_name: str,
         key: str | bytes,
@@ -293,9 +292,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetIfNotExistsResponse:
         """
-        return await self._data_client.set_if_not_exists(cache_name, key, value, ttl)
+        return self._data_client.set_if_not_exists(cache_name, key, value, ttl)
 
-    async def get(self, cache_name: str, key: str | bytes) -> CacheGetResponse:
+    def get(self, cache_name: str, key: str | bytes) -> CacheGetResponse:
         """Get the cache value stored for the given key.
 
         Args:
@@ -305,9 +304,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheGetResponse:
         """
-        return await self._data_client.get(cache_name, key)
+        return self._data_client.get(cache_name, key)
 
-    async def delete(self, cache_name: str, key: str | bytes) -> CacheDeleteResponse:
+    def delete(self, cache_name: str, key: str | bytes) -> CacheDeleteResponse:
         """Remove the key from the cache.
 
         Args:
@@ -317,10 +316,10 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDeleteResponse:
         """
-        return await self._data_client.delete(cache_name, key)
+        return self._data_client.delete(cache_name, key)
 
     # DICTIONARY COLLECTION METHODS
-    async def dictionary_fetch(self, cache_name: str, dictionary_name: str) -> CacheDictionaryFetchResponse:
+    def dictionary_fetch(self, cache_name: str, dictionary_name: str) -> CacheDictionaryFetchResponse:
         """Fetch the entire dictionary from the cache.
 
         Args:
@@ -330,9 +329,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionaryFetchResponse: result of the fetch operation and the associated dictionary.
         """
-        return await self._data_client.dictionary_fetch(cache_name, dictionary_name)
+        return self._data_client.dictionary_fetch(cache_name, dictionary_name)
 
-    async def dictionary_get_field(
+    def dictionary_get_field(
         self, cache_name: str, dictionary_name: str, field: str | bytes
     ) -> CacheDictionaryGetFieldResponse:
         """Get the cache value stored for the given dictionary and field.
@@ -345,7 +344,7 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionaryGetFieldResponse: the status and value for the field.
         """
-        get_fields_response = await self.dictionary_get_fields(cache_name, dictionary_name, [field])
+        get_fields_response = self.dictionary_get_fields(cache_name, dictionary_name, [field])
         if isinstance(get_fields_response, CacheDictionaryGetFields.Hit):
             if len(get_fields_response.responses) == 0:
                 return CacheDictionaryGetField.Error(UnknownException("Unknown get fields response had no data"))
@@ -357,7 +356,7 @@ class SimpleCacheClientAsync:
         else:
             return CacheDictionaryGetField.Error(UnknownException(f"Unknown get field response: {get_fields_response}"))
 
-    async def dictionary_get_fields(
+    def dictionary_get_fields(
         self, cache_name: str, dictionary_name: str, fields: Iterable[str | bytes]
     ) -> CacheDictionaryGetFieldsResponse:
         """Get several values from a dictionary.
@@ -370,9 +369,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionaryGetFieldsResponse: the status and associated value for each field.
         """
-        return await self._data_client.dictionary_get_fields(cache_name, dictionary_name, fields)
+        return self._data_client.dictionary_get_fields(cache_name, dictionary_name, fields)
 
-    async def dictionary_increment(
+    def dictionary_increment(
         self,
         cache_name: str,
         dictionary_name: str,
@@ -395,9 +394,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionaryIncrementResponse: result of the increment operation.
         """
-        return await self._data_client.dictionary_increment(cache_name, dictionary_name, field, amount, ttl)
+        return self._data_client.dictionary_increment(cache_name, dictionary_name, field, amount, ttl)
 
-    async def dictionary_remove_field(
+    def dictionary_remove_field(
         self, cache_name: str, dictionary_name: str, field: str | bytes
     ) -> CacheDictionaryRemoveFieldResponse:
         """Remove a field from a dictionary.
@@ -412,7 +411,7 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionaryRemoveFieldResponse: result of the remove operation.
         """
-        remove_fields_response = await self.dictionary_remove_fields(cache_name, dictionary_name, fields=[field])
+        remove_fields_response = self.dictionary_remove_fields(cache_name, dictionary_name, fields=[field])
         if isinstance(remove_fields_response, CacheDictionaryRemoveFields.Success):
             return CacheDictionaryRemoveField.Success()
         elif isinstance(remove_fields_response, CacheDictionaryRemoveFields.Error):
@@ -422,7 +421,7 @@ class SimpleCacheClientAsync:
                 UnknownException(f"Unknown remove fields response: {remove_fields_response}")
             )
 
-    async def dictionary_remove_fields(
+    def dictionary_remove_fields(
         self, cache_name: str, dictionary_name: str, fields: Iterable[str | bytes]
     ) -> CacheDictionaryRemoveFieldsResponse:
         """Remove fields from a dictionary.
@@ -437,9 +436,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionaryRemoveFieldsResponse: result of the remove fields operation.
         """
-        return await self._data_client.dictionary_remove_fields(cache_name, dictionary_name, fields)
+        return self._data_client.dictionary_remove_fields(cache_name, dictionary_name, fields)
 
-    async def dictionary_set_field(
+    def dictionary_set_field(
         self,
         cache_name: str,
         dictionary_name: str,
@@ -462,9 +461,7 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionarySetFieldResponse: result of the set operation.
         """
-        set_fields_response = await self.dictionary_set_fields(
-            cache_name, dictionary_name, items={field: value}, ttl=ttl
-        )
+        set_fields_response = self.dictionary_set_fields(cache_name, dictionary_name, items={field: value}, ttl=ttl)
         if isinstance(set_fields_response, CacheDictionarySetFields.Success):
             return CacheDictionarySetField.Success()
         elif isinstance(set_fields_response, CacheDictionarySetFields.Error):
@@ -474,7 +471,7 @@ class SimpleCacheClientAsync:
                 UnknownException(f"Unknown set fields response: {set_fields_response}")
             )
 
-    async def dictionary_set_fields(
+    def dictionary_set_fields(
         self,
         cache_name: str,
         dictionary_name: str,
@@ -495,10 +492,10 @@ class SimpleCacheClientAsync:
         Returns:
             CacheDictionarySetFieldsResponse: result of the set fields operation.
         """
-        return await self._data_client.dictionary_set_fields(cache_name, dictionary_name, items, ttl)
+        return self._data_client.dictionary_set_fields(cache_name, dictionary_name, items, ttl)
 
     # LIST COLLECTION METHODS
-    async def list_concatenate_back(
+    def list_concatenate_back(
         self,
         cache_name: str,
         list_name: str,
@@ -520,9 +517,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListConcatenateBackResponse:
         """
-        return await self._data_client.list_concatenate_back(cache_name, list_name, values, ttl, truncate_front_to_size)
+        return self._data_client.list_concatenate_back(cache_name, list_name, values, ttl, truncate_front_to_size)
 
-    async def list_concatenate_front(
+    def list_concatenate_front(
         self,
         cache_name: str,
         list_name: str,
@@ -544,9 +541,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListConcatenateFrontResponse:
         """
-        return await self._data_client.list_concatenate_front(cache_name, list_name, values, ttl, truncate_back_to_size)
+        return self._data_client.list_concatenate_front(cache_name, list_name, values, ttl, truncate_back_to_size)
 
-    async def list_fetch(self, cache_name: str, list_name: str) -> CacheListFetchResponse:
+    def list_fetch(self, cache_name: str, list_name: str) -> CacheListFetchResponse:
         """Gets all values from the list.
 
         Args:
@@ -556,9 +553,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListFetchResponse:
         """
-        return await self._data_client.list_fetch(cache_name, list_name)
+        return self._data_client.list_fetch(cache_name, list_name)
 
-    async def list_length(self, cache_name: str, list_name: str) -> CacheListLengthResponse:
+    def list_length(self, cache_name: str, list_name: str) -> CacheListLengthResponse:
         """Gets the number of values in the list.
 
         Args:
@@ -568,9 +565,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListLengthResponse:
         """
-        return await self._data_client.list_length(cache_name, list_name)
+        return self._data_client.list_length(cache_name, list_name)
 
-    async def list_pop_back(self, cache_name: str, list_name: str) -> CacheListPopBackResponse:
+    def list_pop_back(self, cache_name: str, list_name: str) -> CacheListPopBackResponse:
         """Gets, removes, and returns the last value from the list.
 
         Args:
@@ -580,9 +577,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListPopBackResponse:
         """
-        return await self._data_client.list_pop_back(cache_name, list_name)
+        return self._data_client.list_pop_back(cache_name, list_name)
 
-    async def list_pop_front(self, cache_name: str, list_name: str) -> CacheListPopFrontResponse:
+    def list_pop_front(self, cache_name: str, list_name: str) -> CacheListPopFrontResponse:
         """Gets, removes, and returns the first value from the list.
 
         Args:
@@ -592,9 +589,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListPopFrontResponse:
         """
-        return await self._data_client.list_pop_front(cache_name, list_name)
+        return self._data_client.list_pop_front(cache_name, list_name)
 
-    async def list_push_back(
+    def list_push_back(
         self,
         cache_name: str,
         list_name: str,
@@ -616,9 +613,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListPushBackResponse:
         """
-        return await self._data_client.list_push_back(cache_name, list_name, value, ttl, truncate_front_to_size)
+        return self._data_client.list_push_back(cache_name, list_name, value, ttl, truncate_front_to_size)
 
-    async def list_push_front(
+    def list_push_front(
         self,
         cache_name: str,
         list_name: str,
@@ -640,9 +637,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListPushFrontResponse:
         """
-        return await self._data_client.list_push_front(cache_name, list_name, value, ttl, truncate_back_to_size)
+        return self._data_client.list_push_front(cache_name, list_name, value, ttl, truncate_back_to_size)
 
-    async def list_remove_value(
+    def list_remove_value(
         self,
         cache_name: str,
         list_name: str,
@@ -651,8 +648,8 @@ class SimpleCacheClientAsync:
         """Removes all matching values from the list.
 
         Example:
-            await client.list_concatenate_front(cache_name, list_name, ['up', 'up', 'down', 'down', 'left', 'right'])
-            await client.list_remove_value(cache_name, list_name, 'up')
+            client.list_concatenate_front(cache_name, list_name, ['up', 'up', 'down', 'down', 'left', 'right'])
+            client.list_remove_value(cache_name, list_name, 'up')
             fetch_resp = client.list_fetch(cache_name, list_name)
 
             # ['down', 'down', 'left', 'right']
@@ -666,10 +663,10 @@ class SimpleCacheClientAsync:
         Returns:
             CacheListRemoveValueResponse
         """
-        return await self._data_client.list_remove_value(cache_name, list_name, value)
+        return self._data_client.list_remove_value(cache_name, list_name, value)
 
     # SET COLLECTION METHODS
-    async def set_add_element(
+    def set_add_element(
         self,
         cache_name: str,
         set_name: str,
@@ -688,7 +685,7 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetAddElementResponse
         """
-        resp = await self.set_add_elements(cache_name, set_name, (element,), ttl=ttl)
+        resp = self.set_add_elements(cache_name, set_name, (element,), ttl=ttl)
 
         if isinstance(resp, CacheSetAddElements.Success):
             return CacheSetAddElement.Success()
@@ -697,7 +694,7 @@ class SimpleCacheClientAsync:
         else:
             raise UnknownException(f"Unknown response type: {type(resp)}")
 
-    async def set_add_elements(
+    def set_add_elements(
         self,
         cache_name: str,
         set_name: str,
@@ -716,9 +713,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetAddElementsResponse
         """
-        return await self._data_client.set_add_elements(cache_name, set_name, elements, ttl)
+        return self._data_client.set_add_elements(cache_name, set_name, elements, ttl)
 
-    async def set_fetch(
+    def set_fetch(
         self,
         cache_name: str,
         set_name: str,
@@ -732,11 +729,9 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetFetchResponse
         """
-        return await self._data_client.set_fetch(cache_name, set_name)
+        return self._data_client.set_fetch(cache_name, set_name)
 
-    async def set_remove_element(
-        self, cache_name: str, set_name: str, element: str | bytes
-    ) -> CacheSetRemoveElementResponse:
+    def set_remove_element(self, cache_name: str, set_name: str, element: str | bytes) -> CacheSetRemoveElementResponse:
         """Remove an element from a set.
 
         Args:
@@ -747,7 +742,7 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetRemoveElementResponse
         """
-        resp = await self.set_remove_elements(cache_name, set_name, (element,))
+        resp = self.set_remove_elements(cache_name, set_name, (element,))
         if isinstance(resp, CacheSetRemoveElements.Success):
             return CacheSetRemoveElement.Success()
         elif isinstance(resp, CacheSetRemoveElements.Error):
@@ -755,7 +750,7 @@ class SimpleCacheClientAsync:
         else:
             raise UnknownException(f"Unknown response type: {type(resp)}")
 
-    async def set_remove_elements(
+    def set_remove_elements(
         self, cache_name: str, set_name: str, elements: Iterable[str | bytes]
     ) -> CacheSetRemoveElementsResponse:
         """Remove elements from a set.
@@ -768,7 +763,7 @@ class SimpleCacheClientAsync:
         Returns:
             CacheSetRemoveElementsResponse
         """
-        return await self._data_client.set_remove_elements(cache_name, set_name, elements)
+        return self._data_client.set_remove_elements(cache_name, set_name, elements)
 
     @property
     def _data_client(self) -> _ScsDataClient:
