+++ src/momento/internal/synchronous/_scs_grpc_manager.py
@@ -20,7 +20,7 @@ class _ControlGrpcManager:
 
     version = pkg_resources.get_distribution("momento").version
 
-    def __init__(self, credential_provider: CredentialProvider, configuration: Configuration):
+    def __init__(self, configuration: Configuration, credential_provider: CredentialProvider):
         self._secure_channel = grpc.secure_channel(
             target=credential_provider.control_endpoint, credentials=grpc.ssl_channel_credentials()
         )
