+++ tests/momento/simple_cache_client/test_dictionary_async.py
@@ -9,7 +9,7 @@ from pytest import fixture
 from pytest_describe import behaves_like
 from typing_extensions import Protocol
 
-from momento import SimpleCacheClientAsync
+from momento import CacheClientAsync
 from momento.auth import CredentialProvider
 from momento.config import Configuration
 from momento.errors import MomentoErrorCode
@@ -134,7 +134,7 @@ class TDictionaryGetter(Protocol):
 def a_dictionary_getter() -> None:
     async def with_bytes_it_succeeds(
         dictionary_getter: TDictionaryGetter,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_bytes: bytes,
@@ -154,7 +154,7 @@ def a_dictionary_getter() -> None:
 
     async def with_string_it_succeeds(
         dictionary_getter: TDictionaryGetter,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -182,7 +182,7 @@ class TDictionaryRemover(Protocol):
 
 def a_dictionary_remover() -> None:
     async def with_string_field_succeeds(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -204,7 +204,7 @@ def a_dictionary_remover() -> None:
 
     async def with_bytes_field_succeeds(
         dictionary_remover: TDictionaryRemover,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_bytes: bytes,
@@ -227,7 +227,7 @@ def a_dictionary_remover() -> None:
 class TDictionarySetter(Protocol):
     def __call__(
         self,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         field: TDictionaryField,
@@ -247,7 +247,7 @@ def a_dictionary_setter() -> None:
         dictionary_name: TDictionaryName,
         dictionary_items: TDictionaryItems,
     ) -> None:
-        async with SimpleCacheClientAsync(configuration, credential_provider, timedelta(hours=1)) as client:
+        async with CacheClientAsync(configuration, credential_provider, timedelta(hours=1)) as client:
             ttl_seconds = 0.5
             ttl = CollectionTtl(ttl=timedelta(seconds=ttl_seconds), refresh_ttl=False)
 
@@ -260,7 +260,7 @@ def a_dictionary_setter() -> None:
             assert isinstance(fetch_response, CacheDictionaryFetch.Miss)
 
     async def it_refreshes_the_ttl(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         dictionary_setter: TDictionarySetter,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
@@ -287,7 +287,7 @@ def a_dictionary_setter() -> None:
         dictionary_value_str: str,
     ) -> None:
         ttl_seconds = 1
-        async with SimpleCacheClientAsync(configuration, credential_provider, timedelta(seconds=ttl_seconds)) as client:
+        async with CacheClientAsync(configuration, credential_provider, timedelta(seconds=ttl_seconds)) as client:
             ttl = CollectionTtl.from_cache_ttl().with_no_refresh_ttl_on_updates()
 
             await dictionary_setter(
@@ -306,7 +306,7 @@ def a_dictionary_setter() -> None:
 
     async def with_bytes_it_succeeds(
         dictionary_setter: TDictionarySetter,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_bytes: bytes,
@@ -325,7 +325,7 @@ def a_dictionary_setter() -> None:
 
     async def with_strings_it_succeeds(
         dictionary_setter: TDictionarySetter,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -338,7 +338,7 @@ def a_dictionary_setter() -> None:
 
     async def with_other_values_type_it_errors(
         dictionary_setter: TDictionarySetter,
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         dictionary_name: TDictionaryName,
     ) -> None:
         for field, value, bad_type in [
@@ -385,7 +385,7 @@ def a_dictionary_value_validator() -> None:
 def describe_dictionary_get_field() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync, dictionary_name: TDictionaryName, dictionary_field_str: str
+        client_async: CacheClientAsync, dictionary_name: TDictionaryName, dictionary_field_str: str
     ) -> TCacheNameValidator:
         return partial(client_async.dictionary_get_field, dictionary_name=dictionary_name, field=dictionary_field_str)
 
@@ -393,7 +393,7 @@ def describe_dictionary_get_field() -> None:
     def connection_validator(
         cache_name: TCacheName, dictionary_name: TDictionaryName, dictionary_field_str: str
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_get_field(
                 cache_name, dictionary_name=dictionary_name, field=dictionary_field_str
             )
@@ -402,7 +402,7 @@ def describe_dictionary_get_field() -> None:
 
     @fixture
     def dictionary_field_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
     ) -> TDictionaryFieldValidator:
@@ -410,18 +410,18 @@ def describe_dictionary_get_field() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_field_str: str,
     ) -> TDictionaryNameValidator:
         return partial(client_async.dictionary_get_field, cache_name=cache_name, field=dictionary_field_str)
 
     @fixture
-    def dictionary_getter(client_async: SimpleCacheClientAsync) -> TDictionaryGetter:
+    def dictionary_getter(client_async: CacheClientAsync) -> TDictionaryGetter:
         return partial(client_async.dictionary_get_field)
 
     async def misses_when_the_dictionary_does_not_exist(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -438,7 +438,7 @@ def describe_dictionary_get_field() -> None:
 def describe_dictionary_get_fields() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync, dictionary_name: TDictionaryName, dictionary_fields: TDictionaryFields
+        client_async: CacheClientAsync, dictionary_name: TDictionaryName, dictionary_fields: TDictionaryFields
     ) -> TCacheNameValidator:
         return partial(client_async.dictionary_get_fields, dictionary_name=dictionary_name, fields=dictionary_fields)
 
@@ -446,14 +446,14 @@ def describe_dictionary_get_fields() -> None:
     def connection_validator(
         cache_name: TCacheName, dictionary_name: TDictionaryName, dictionary_fields: TDictionaryFields
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_get_fields(cache_name, dictionary_name, dictionary_fields)
 
         return _connection_validator
 
     @fixture
     def dictionary_fields_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
     ) -> TDictionaryFieldsValidator:
@@ -461,14 +461,14 @@ def describe_dictionary_get_fields() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_field_str: str,
     ) -> TDictionaryNameValidator:
         return partial(client_async.dictionary_get_fields, cache_name=cache_name, fields=dictionary_field_str)
 
     @fixture
-    def dictionary_getter(client_async: SimpleCacheClientAsync) -> TDictionaryGetter:
+    def dictionary_getter(client_async: CacheClientAsync) -> TDictionaryGetter:
         async def _dictionary_getter(
             cache_name: TCacheName, dictionary_name: TDictionaryName, field: TDictionaryField
         ) -> CacheDictionaryGetFieldResponse:
@@ -483,7 +483,7 @@ def describe_dictionary_get_fields() -> None:
         return _dictionary_getter
 
     async def it_misses_when_the_dictionary_does_not_exist(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_fields: TDictionaryFields,
@@ -492,7 +492,7 @@ def describe_dictionary_get_fields() -> None:
         assert isinstance(get_response, CacheDictionaryGetFields.Miss)
 
     async def it_decodes_responses_to_dictionaries_correctly(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
     ) -> None:
@@ -515,7 +515,7 @@ def describe_dictionary_get_fields() -> None:
         assert get_response.value_dictionary_string_string == dictionary_items
 
     async def it_excludes_misses_from_dictionary(
-        client_async: SimpleCacheClientAsync, cache_name: TCacheName, dictionary_name: TDictionaryName
+        client_async: CacheClientAsync, cache_name: TCacheName, dictionary_name: TDictionaryName
     ) -> None:
         dictionary_items: dict[str, str] = dict([(uuid_str(), uuid_str()), (uuid_str(), uuid_str())])
         set_response = await client_async.dictionary_set_fields(cache_name, dictionary_name, dictionary_items)
@@ -540,26 +540,22 @@ def describe_dictionary_get_fields() -> None:
 @behaves_like(a_dictionary_name_validator)
 def describe_dictionary_fetch() -> None:
     @fixture
-    def cache_name_validator(
-        client_async: SimpleCacheClientAsync, dictionary_name: TDictionaryName
-    ) -> TCacheNameValidator:
+    def cache_name_validator(client_async: CacheClientAsync, dictionary_name: TDictionaryName) -> TCacheNameValidator:
         return partial(client_async.dictionary_fetch, dictionary_name=dictionary_name)
 
     @fixture
     def connection_validator(cache_name: TCacheName, dictionary_name: TDictionaryName) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_fetch(cache_name, dictionary_name=dictionary_name)
 
         return _connection_validator
 
     @fixture
-    def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync, cache_name: TCacheName
-    ) -> TDictionaryNameValidator:
+    def dictionary_name_validator(client_async: CacheClientAsync, cache_name: TCacheName) -> TDictionaryNameValidator:
         return partial(client_async.dictionary_fetch, cache_name=cache_name)
 
     async def misses_when_the_dictionary_does_not_exist(
-        client_async: SimpleCacheClientAsync, cache_name: TCacheName, dictionary_name: TDictionaryName
+        client_async: CacheClientAsync, cache_name: TCacheName, dictionary_name: TDictionaryName
     ) -> None:
         fetch_response = await client_async.dictionary_fetch(cache_name, dictionary_name)
         assert isinstance(fetch_response, CacheDictionaryFetch.Miss)
@@ -572,7 +568,7 @@ def describe_dictionary_fetch() -> None:
 def describe_dictionary_increment() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
         increment_amount: int,
@@ -591,7 +587,7 @@ def describe_dictionary_increment() -> None:
         dictionary_field_str: str,
         increment_amount: int,
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_increment(
                 cache_name, dictionary_name=dictionary_name, field=dictionary_field_str, amount=increment_amount
             )
@@ -600,7 +596,7 @@ def describe_dictionary_increment() -> None:
 
     @fixture
     def dictionary_field_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         increment_amount: int,
@@ -614,7 +610,7 @@ def describe_dictionary_increment() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_field_str: str,
         increment_amount: int,
@@ -627,7 +623,7 @@ def describe_dictionary_increment() -> None:
         )
 
     async def it_starts_at_zero(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -639,7 +635,7 @@ def describe_dictionary_increment() -> None:
         assert increment_response.value == 0
 
     async def it_increments_when_the_field_does_not_exist(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -652,7 +648,7 @@ def describe_dictionary_increment() -> None:
         assert increment_response.value == increment_amount
 
     async def it_increments_an_existing_field(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -670,7 +666,7 @@ def describe_dictionary_increment() -> None:
         assert increment_response.value == 2 * increment_amount
 
     async def it_errors_on_bad_initial_values(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -696,7 +692,7 @@ def describe_dictionary_increment() -> None:
 def describe_dictionary_remove_field() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync, dictionary_name: TDictionaryName, dictionary_field_str: str
+        client_async: CacheClientAsync, dictionary_name: TDictionaryName, dictionary_field_str: str
     ) -> TCacheNameValidator:
         return partial(
             client_async.dictionary_remove_field, dictionary_name=dictionary_name, field=dictionary_field_str
@@ -706,7 +702,7 @@ def describe_dictionary_remove_field() -> None:
     def connection_validator(
         cache_name: TCacheName, dictionary_name: TDictionaryName, dictionary_field_str: str
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_remove_field(
                 cache_name, dictionary_name=dictionary_name, field=dictionary_field_str
             )
@@ -715,7 +711,7 @@ def describe_dictionary_remove_field() -> None:
 
     @fixture
     def dictionary_field_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
     ) -> TDictionaryFieldValidator:
@@ -723,18 +719,18 @@ def describe_dictionary_remove_field() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_field_str: str,
     ) -> TDictionaryNameValidator:
         return partial(client_async.dictionary_remove_field, cache_name=cache_name, field=dictionary_field_str)
 
     @fixture
-    def dictionary_remover(client_async: SimpleCacheClientAsync) -> TDictionaryRemover:
+    def dictionary_remover(client_async: CacheClientAsync) -> TDictionaryRemover:
         return partial(client_async.dictionary_remove_field)
 
     async def it_removes_a_field(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -769,7 +765,7 @@ def describe_dictionary_remove_field() -> None:
 def describe_dictionary_remove_fields() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         dictionary_name: TDictionaryName,
         dictionary_fields: TDictionaryFields,
     ) -> TCacheNameValidator:
@@ -779,7 +775,7 @@ def describe_dictionary_remove_fields() -> None:
     def connection_validator(
         cache_name: TCacheName, dictionary_name: TDictionaryName, dictionary_fields: TDictionaryFields
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_remove_fields(
                 cache_name, dictionary_name=dictionary_name, fields=dictionary_fields
             )
@@ -788,7 +784,7 @@ def describe_dictionary_remove_fields() -> None:
 
     @fixture
     def dictionary_fields_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
     ) -> TDictionaryFieldsValidator:
@@ -796,14 +792,14 @@ def describe_dictionary_remove_fields() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_fields: TDictionaryFields,
     ) -> TDictionaryNameValidator:
         return partial(client_async.dictionary_remove_fields, cache_name=cache_name, fields=dictionary_fields)
 
     @fixture
-    def dictionary_remover(client_async: SimpleCacheClientAsync) -> TDictionaryRemover:
+    def dictionary_remover(client_async: CacheClientAsync) -> TDictionaryRemover:
         async def _dictionary_remover(
             cache_name: TCacheName, dictionary_name: TDictionaryName, field: TDictionaryField
         ) -> CacheResponse:
@@ -814,7 +810,7 @@ def describe_dictionary_remove_fields() -> None:
         return _dictionary_remover
 
     async def it_removes_multiple_items(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_items: TDictionaryItems,
@@ -845,7 +841,7 @@ def describe_dictionary_remove_fields() -> None:
 def describe_dictionary_set_field() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
         dictionary_value_str: str,
@@ -861,7 +857,7 @@ def describe_dictionary_set_field() -> None:
     def connection_validator(
         cache_name: TCacheName, dictionary_name: TDictionaryName, dictionary_field_str: str, dictionary_value_str: str
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_set_field(
                 cache_name,
                 dictionary_name=dictionary_name,
@@ -873,7 +869,7 @@ def describe_dictionary_set_field() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_field_str: str,
         dictionary_value_str: str,
@@ -888,7 +884,7 @@ def describe_dictionary_set_field() -> None:
     @fixture
     def dictionary_setter() -> TDictionarySetter:
         async def _dictionary_setter(
-            client_async: SimpleCacheClientAsync,
+            client_async: CacheClientAsync,
             cache_name: TCacheName,
             dictionary_name: TDictionaryName,
             field: TDictionaryField,
@@ -908,7 +904,7 @@ def describe_dictionary_set_field() -> None:
 
     @fixture
     def dictionary_value_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_field_str: str,
@@ -929,7 +925,7 @@ def describe_dictionary_set_field() -> None:
 def describe_dictionary_set_fields() -> None:
     @fixture
     def cache_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         dictionary_name: TDictionaryName,
         dictionary_items: TDictionaryItems,
     ) -> TCacheNameValidator:
@@ -939,7 +935,7 @@ def describe_dictionary_set_fields() -> None:
     def connection_validator(
         cache_name: TCacheName, dictionary_name: TDictionaryName, dictionary_items: TDictionaryItems
     ) -> TConnectionValidator:
-        async def _connection_validator(client_async: SimpleCacheClientAsync) -> CacheResponse:
+        async def _connection_validator(client_async: CacheClientAsync) -> CacheResponse:
             return await client_async.dictionary_set_fields(
                 cache_name, dictionary_name=dictionary_name, items=dictionary_items
             )
@@ -948,7 +944,7 @@ def describe_dictionary_set_fields() -> None:
 
     @fixture
     def dictionary_items_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
     ) -> TDictionaryItemsValidator:
@@ -960,7 +956,7 @@ def describe_dictionary_set_fields() -> None:
 
     @fixture
     def dictionary_name_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_items: TDictionaryItems,
     ) -> TDictionaryNameValidator:
@@ -969,7 +965,7 @@ def describe_dictionary_set_fields() -> None:
     @fixture
     def dictionary_setter() -> TDictionarySetter:
         async def _dictionary_setter(
-            client_async: SimpleCacheClientAsync,
+            client_async: CacheClientAsync,
             cache_name: TCacheName,
             dictionary_name: TDictionaryName,
             field: TDictionaryField,
@@ -988,7 +984,7 @@ def describe_dictionary_set_fields() -> None:
 
     @fixture
     def dictionary_value_validator(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         field: TDictionaryField,
@@ -1003,7 +999,7 @@ def describe_dictionary_set_fields() -> None:
         return _value_validator
 
     async def it_sets_multiple_items(
-        client_async: SimpleCacheClientAsync,
+        client_async: CacheClientAsync,
         cache_name: TCacheName,
         dictionary_name: TDictionaryName,
         dictionary_items: TDictionaryItems,
